apps/django_backend/requirements/base.txt
pytz==2023.3  # https://github.com/stub42/pytz
python-slugify==8.0.1  # https://github.com/un33k/python-slugify
Pillow==10.0.0  # https://github.com/python-pillow/Pillow
argon2-cffi==23.1.0  # https://github.com/hynek/argon2_cffi
redis==5.0.0  # https://github.com/redis/redis-py
hiredis==2.2.3  # https://github.com/redis/hiredis-py
celery==5.3.4  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.5.0  # https://github.com/celery/django-celery-beat
flower==2.0.1  # https://github.com/mher/flower
uvicorn[standard]==0.23.2  # https://github.com/encode/uvicorn
httpx==0.24.1
xxhash~=3.0
validators==0.22.0

# Other
# ------------------------------------------------------------------------------
cachetools==5.3.1
# Django
# ------------------------------------------------------------------------------
django==4.2.5  # pyup: < 4.1  # https://www.djangoproject.com/
django-environ==0.10.0  # https://github.com/joke2k/django-environ
django-filter==23.2
django-model-utils==4.3.1  # https://github.com/jazzband/django-model-utils
dj-rest-auth==4.0.1
djangorestframework-simplejwt==5.3.0
django-redis==5.3.0  # https://github.com/jazzband/django-redis
django-resized~=1.0
django-countries==7.5.1
django-import-export==3.2.0
# Django REST Framework
djangorestframework==3.14.0  # https://github.com/encode/django-rest-framework
django-cors-headers==4.2.0  # https://github.com/adamchainz/django-cors-headers
# DRF-spectacular for api documentation
drf-spectacular==0.26.4  # https://github.com/tfranzel/drf-spectacular
